requirements.txt
# Core application libraries
langchain
google-generativeai
pydantic
python-dotenv
rich
streamlit
requests 
Flask

# RAG libraries
faiss-cpu
langchain-huggingface>=0.0.3

# LangChain Core Components
langchain-core>=1.0.0
langchain-community>=0.0.30
langchain-google-genai>=1.0.0
